refactor: move homepage navigation bar into homepage page
